pyproject.toml
[tool.poetry]
name = "rulesofriveravulsion"
version = "0.1.0"
description = "reproducibility package for rules of river avulsion change downstream manuscript"
authors = ["Jake Gearon <jake.gearon@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
geopandas = "^0.14.1"
pyqt5 = "^5.15.10"
scipy = "^1.11.4"
scikit-image = "^0.22.0"
rasterio = "^1.3.9"
shapely = "^2.0.2"
pysheds = "^0.3.5"
plotly-express = "^0.4.1"
pyarrow = "^14.0.1"
nbformat = "^5.9.2"
earthengine-api = "^0.1.383"
geemap = "^0.29.6"
pycrs = "^1.0.2"
seaborn = "^0.13.0"
xgboost = "^2.0.2"
scikit-learn = "^1.3.2"
openpyxl = "^3.1.2"
scikit-optimize = {git = "https://github.com/mikesmithlab/scikit-optimize.git"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
